requirements.txt
# pip install -r requirements.txt

# numpy==1.23.1
# pandas==1.4.3
# scipy==1.9.0
# tqdm==4.64.0
# # wandb==0.13
# sklearn==0.0
# pytorch_forecasting==0.10.3
# snntorch==0.5.3

pandas
scipy
tqdm
scikit-learn
matplotlib
wandb
snntorch
wget